requirements.txt
allure-python-commons~=2.13.5
requests~=2.32.3
pytest~=8.3.5
allure-pytest==2.13.5